Updated for resubmission the app.js file
After grading it needed a method to check if an invalid answer(no selection) had been made on the radio buttons when submitted.  Added an invalidAnswer variable to the store, and a conditional statement to check if invalid answer is true then display "Please select a valid answer" on screen.  Now users cannot skip questions by submitting a blank answer.